create a root user with environment variables and put module in its own file
